TextInput -> Input and Fixed BlockBool
