+++ src/app/admin/pageform/blockeditor/blockform/blockarray/BlockArray.jsx
@@ -2,7 +2,7 @@ import React from 'react'
 import shortid from 'shortid'
 
 import DeleteBtn from 'blockform/modules/deletebtn/DeleteBtn.jsx'
-import TextInput from 'modules/textinput/TextInput.jsx'
+import Input from 'modules/input/Input.jsx'
 import styles from './blockArray.mod.scss'
 
 export default function BlockText(props) {
@@ -29,7 +29,7 @@ export default function BlockText(props) {
   return(
     <div className={styles.wrapper} onClick={focus}>
       <span>{props.name}: </span>
-        <TextInput
+        <Input
           name={props.name}
           text={content()}
           onChange={onChange}
