+++ src/app/admin/pageform/blockeditor/blockform/blockbool/BlockBool.jsx
@@ -2,25 +2,23 @@ import React from 'react'
 import shortid from 'shortid'
 
 import DeleteBtn from 'blockform/modules/deletebtn/DeleteBtn.jsx'
-import TextInput from 'modules/textinput/TextInput.jsx'
+import Input from 'modules/input/Input.jsx'
 import styles from './blockBool.mod.scss'
 
 export default function BlockText(props) {
   const [id] = React.useState(shortid.generate())
 
-  function onChange(event) {
-    if (event && event.target) {
-      props.onChange(event.target.checked, props.name)
-    }
+  function onChange(value) {
+    console.log(value)
+    props.onChange(value, props.name)
   }
 
   return(
     <div className={styles.wrapper} >
       <span>{props.name}: </span>
-        <TextInput
-          name={props.name}
+        <Input
           checked={props.content}
-          onChange={props.onChange}
+          onChange={onChange}
           type={'checkbox'}
           classNames={styles.text} />
