+++ src/app/admin/pageform/blockeditor/blockform/blockbool/blockBool.mod.scss
@@ -10,6 +10,7 @@
   }
 
   .text {
+    height: 100%;
     border: none
   }
 }
