+++ src/app/admin/pageform/blockeditor/blockform/blocktext/BlockText.jsx
@@ -2,7 +2,7 @@ import React from 'react'
 import shortid from 'shortid'
 
 import DeleteBtn from 'blockform/modules/deletebtn/DeleteBtn.jsx'
-import TextInput from 'modules/textinput/TextInput.jsx'
+import Input from 'modules/input/Input.jsx'
 import styles from './blockText.mod.scss'
 
 export default function BlockNumber(props) {
@@ -19,7 +19,7 @@ export default function BlockNumber(props) {
   return(
     <div className={styles.wrapper} onClick={focus}>
       <span>{props.name}: </span>
-      <TextInput
+      <Input
         name={props.name}
         text={props.content}
         onChange={props.onChange}
