+++ src/app/admin/pageform/blockeditor/blockform/modules/addlist/AddList.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 
 import ToggleList from 'modules/lists/togglelist/ToggleList.jsx'
-import TextInput from 'modules/textinput/TextInput.jsx'
+import Input from 'modules/input/Input.jsx'
 
 import ArrowImg from 'public/images/arrow-down.png';
 import AddButton from 'public/images/add-button.png'
@@ -29,7 +29,7 @@ export default function AddList(props) {
     <div className={styles.wrapper}>
       <div className={styles.header}>
         <img className={styles.arrow} src={ArrowImg} onClick={toggleList} ref={imgRef}/>
-        <TextInput
+        <Input
           text={custom}
           onChange={setCustom}
           classNames={styles.custom} />
