+++ src/app/admin/pageform/header/Header.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 
-import TextInput from 'modules/textinput/TextInput.jsx'
+import Input from 'modules/input/Input.jsx'
 
 import styles from './header.mod.scss'
 
@@ -8,37 +8,37 @@ export default function Header(props) {
 
   return(
     <div className={styles.headerWrapper}>
-      <TextInput classNames={styles.headerText}
+      <Input classNames={styles.headerText}
         text={props.header}
         name="header"
         onChange={props.onChange}>
 
         <span className={styles.headerLabel}>Header</span>
-      </TextInput>
+      </Input>
 
-      <TextInput classNames={styles.sequenceText}
+      <Input classNames={styles.sequenceText}
         text={props.sequence}
         name="sequence"
         onChange={props.onChange}>
 
         <span className={styles.sequenceLabel}>Sequence</span>
-      </TextInput>
+      </Input>
 
-      <TextInput classNames={styles.captionText}
+      <Input classNames={styles.captionText}
         text={props.caption}
         name="caption"
         onChange={props.onChange}>
 
         <span className={styles.captionLabel}>Caption</span>
-      </TextInput>
+      </Input>
 
-      <TextInput classNames={styles.slugText}
+      <Input classNames={styles.slugText}
         text={props.slug}
         name="slug"
         onChange={props.onChange}>
 
         <span className={styles.slugLabel}>Slug</span>
-      </TextInput>
+      </Input>
     </div>
   )
 }
